Squashed commit of the following:
commit 712fb69285e0e04342aefcc58b7a232ce4f48b59
Author: AJ Slater <aj@slater.net>
Date:   Mon Jul 11 13:55:36 2022 -0700

    NEWS

commit fb518bc8f4d0cf1bf24194a8e1268941b1c18e78
Author: AJ Slater <aj@slater.net>
Date:   Mon Jul 11 13:53:16 2022 -0700

    block all of title bar with isOpenToSee

commit 12265a74caa3ef3a5dc818e7e87c93c22d8187b6
Author: AJ Slater <aj@slater.net>
Date:   Mon Jul 11 13:48:38 2022 -0700

    removed clever panel resuse in favor of a more direct method
